WIP: Change methods to inject sub stores for account store
