+++ packages/stores/src/account/base.ts
@@ -1,36 +1,8 @@
-import {
-  action,
-  computed,
-  flow,
-  makeObservable,
-  observable,
-  runInAction,
-} from "mobx";
+import { action, computed, flow, makeObservable, observable } from "mobx";
 import { AppCurrency, Keplr, KeplrSignOptions } from "@keplr-wallet/types";
-import { DeepReadonly } from "utility-types";
-import { ChainGetter, IObject } from "../common";
-import { IQueriesStore, QueriesSetBase } from "../query";
+import { ChainGetter } from "../common";
 import { DenomHelper, toGenerator } from "@keplr-wallet/common";
-import {
-  BroadcastMode,
-  makeSignDoc,
-  makeStdTx,
-  Msg,
-  StdFee,
-  StdSignDoc,
-} from "@cosmjs/launchpad";
-import {
-  BaseAccount,
-  cosmos,
-  google,
-  TendermintTxTracer,
-} from "@keplr-wallet/cosmos";
-import Axios, { AxiosInstance } from "axios";
-import { Buffer } from "buffer/";
-import Long from "long";
-import ICoin = cosmos.base.v1beta1.ICoin;
-import SignMode = cosmos.tx.signing.v1beta1.SignMode;
-
+import { StdFee } from "@cosmjs/launchpad";
 import { evmosToEth } from "@hanchon/ethermint-address-converter";
 
 export enum WalletStatus {
@@ -46,18 +18,7 @@ export interface MsgOpt {
   readonly gas: number;
 }
 
-/*
-  If the chain has "no-legacy-stdTx" feature, we should send the tx based on protobuf.
-  Expectedly, the sign doc should be formed as animo-json regardless of the tx type (animo or proto).
-*/
-type AminoMsgsOrWithProtoMsgs =
-  | Msg[]
-  | {
-      aminoMsgs: Msg[];
-      protoMsgs?: google.protobuf.IAny[];
-    };
-
-export interface AccountSetOpts<MsgOpts> {
+export interface AccountSetOpts {
   readonly prefetching: boolean;
   readonly suggestChain: boolean;
   readonly suggestChainFn?: (
@@ -65,20 +26,10 @@ export interface AccountSetOpts<MsgOpts> {
     chainInfo: ReturnType<ChainGetter["getChain"]>
   ) => Promise<void>;
   readonly autoInit: boolean;
-  readonly preTxEvents?: {
-    onBroadcastFailed?: (e?: Error) => void;
-    onBroadcasted?: (txHash: Uint8Array) => void;
-    onFulfill?: (tx: any) => void;
-  };
   readonly getKeplr: () => Promise<Keplr | undefined>;
-  readonly msgOpts: MsgOpts;
-  readonly wsObject?: new (
-    url: string,
-    protocols?: string | string[]
-  ) => WebSocket;
 }
 
-export class AccountSetBase<MsgOpts, Queries extends IObject> {
+export class AccountSetBase {
   @observable
   protected _walletVersion: string | undefined = undefined;
 
@@ -92,9 +43,7 @@ export class AccountSetBase<MsgOpts, Queries extends IObject> {
   protected _bech32Address: string = "";
 
   @observable
-  protected _isSendingMsg: string | boolean = false;
-
-  public broadcastMode: "sync" | "async" | "block" = "sync";
+  protected _txTypeInProgress: string = "";
 
   protected pubKey: Uint8Array;
 
@@ -123,8 +72,7 @@ export class AccountSetBase<MsgOpts, Queries extends IObject> {
     },
     protected readonly chainGetter: ChainGetter,
     protected readonly chainId: string,
-    protected readonly queriesStore: IQueriesStore<Queries>,
-    protected readonly opts: AccountSetOpts<MsgOpts>
+    protected readonly opts: AccountSetOpts
   ) {
     makeObservable(this);
 
@@ -137,10 +85,6 @@ export class AccountSetBase<MsgOpts, Queries extends IObject> {
 
   getKeplr(): Promise<Keplr | undefined> {
     return this.opts.getKeplr();
-  }
-
-  get msgOpts(): MsgOpts {
-    return this.opts.msgOpts;
   }
 
   registerSendTokenFn(
@@ -247,127 +191,20 @@ export class AccountSetBase<MsgOpts, Queries extends IObject> {
   }
 
   @computed
-  get isReadyToSendMsgs(): boolean {
+  get isReadyToSendTx(): boolean {
     return (
       this.walletStatus === WalletStatus.Loaded && this.bech32Address !== ""
     );
   }
 
-  async sendMsgs(
-    type: string | "unknown",
-    msgs:
-      | AminoMsgsOrWithProtoMsgs
-      | (() => Promise<AminoMsgsOrWithProtoMsgs> | AminoMsgsOrWithProtoMsgs),
-    memo: string = "",
-    fee: StdFee,
-    signOptions?: KeplrSignOptions,
-    onTxEvents?:
-      | ((tx: any) => void)
-      | {
-          onBroadcastFailed?: (e?: Error) => void;
-          onBroadcasted?: (txHash: Uint8Array) => void;
-          onFulfill?: (tx: any) => void;
-        }
-  ) {
-    runInAction(() => {
-      this._isSendingMsg = type;
-    });
-
-    let txHash: Uint8Array;
-    let signDoc: StdSignDoc;
-    try {
-      if (typeof msgs === "function") {
-        msgs = await msgs();
-      }
-
-      const result = await this.broadcastMsgs(
-        msgs,
-        fee,
-        memo,
-        signOptions,
-        this.broadcastMode
-      );
-      txHash = result.txHash;
-      signDoc = result.signDoc;
-    } catch (e) {
-      runInAction(() => {
-        this._isSendingMsg = false;
-      });
-
-      if (this.opts.preTxEvents?.onBroadcastFailed) {
-        this.opts.preTxEvents.onBroadcastFailed(e);
-      }
-
-      if (
-        onTxEvents &&
-        "onBroadcastFailed" in onTxEvents &&
-        onTxEvents.onBroadcastFailed
-      ) {
-        onTxEvents.onBroadcastFailed(e);
-      }
-
-      throw e;
-    }
-
-    let onBroadcasted: ((txHash: Uint8Array) => void) | undefined;
-    let onFulfill: ((tx: any) => void) | undefined;
-
-    if (onTxEvents) {
-      if (typeof onTxEvents === "function") {
-        onFulfill = onTxEvents;
-      } else {
-        onBroadcasted = onTxEvents.onBroadcasted;
-        onFulfill = onTxEvents.onFulfill;
-      }
-    }
-
-    if (this.opts.preTxEvents?.onBroadcasted) {
-      this.opts.preTxEvents.onBroadcasted(txHash);
-    }
-    if (onBroadcasted) {
-      onBroadcasted(txHash);
-    }
-
-    const txTracer = new TendermintTxTracer(
-      this.chainGetter.getChain(this.chainId).rpc,
-      "/websocket",
-      {
-        wsObject: this.opts.wsObject,
-      }
+  /**
+   * @deprecated Use `isReadyToSendTx`
+   */
+  @computed
+  get isReadyToSendMsgs(): boolean {
+    return (
+      this.walletStatus === WalletStatus.Loaded && this.bech32Address !== ""
     );
-    txTracer.traceTx(txHash).then((tx) => {
-      txTracer.close();
-
-      runInAction(() => {
-        this._isSendingMsg = false;
-      });
-
-      // After sending tx, the balances is probably changed due to the fee.
-      for (const feeAmount of signDoc.fee.amount) {
-        const bal = this.queries.queryBalances
-          .getQueryBech32Address(this.bech32Address)
-          .balances.find(
-            (bal) => bal.currency.coinMinimalDenom === feeAmount.denom
-          );
-
-        if (bal) {
-          bal.fetch();
-        }
-      }
-
-      // Always add the tx hash data.
-      if (tx && !tx.hash) {
-        tx.hash = Buffer.from(txHash).toString("hex");
-      }
-
-      if (this.opts.preTxEvents?.onFulfill) {
-        this.opts.preTxEvents.onFulfill(tx);
-      }
-
-      if (onFulfill) {
-        onFulfill(tx);
-      }
-    });
   }
 
   async sendToken(
@@ -407,124 +244,6 @@ export class AccountSetBase<MsgOpts, Queries extends IObject> {
     throw new Error(`Unsupported type of currency (${denomHelper.type})`);
   }
 
-  // Return the tx hash.
-  protected async broadcastMsgs(
-    msgs: AminoMsgsOrWithProtoMsgs,
-    fee: StdFee,
-    memo: string = "",
-    signOptions?: KeplrSignOptions,
-    mode: "block" | "async" | "sync" = "async"
-  ): Promise<{
-    txHash: Uint8Array;
-    signDoc: StdSignDoc;
-  }> {
-    if (this.walletStatus !== WalletStatus.Loaded) {
-      throw new Error(`Wallet is not loaded: ${this.walletStatus}`);
-    }
-
-    let aminoMsgs: Msg[];
-    let protoMsgs: google.protobuf.IAny[] | undefined;
-    if ("aminoMsgs" in msgs) {
-      aminoMsgs = msgs.aminoMsgs;
-      protoMsgs = msgs.protoMsgs;
-    } else {
-      aminoMsgs = msgs;
-    }
-
-    if (aminoMsgs.length === 0) {
-      throw new Error("There is no msg to send");
-    }
-
-    if (
-      this.hasNoLegacyStdFeature() &&
-      (!protoMsgs || protoMsgs.length === 0)
-    ) {
-      throw new Error(
-        "Chain can't send legecy stdTx. But, proto any type msgs are not provided"
-      );
-    }
-
-    const account = await BaseAccount.fetchFromRest(
-      this.instance,
-      this.bech32Address,
-      true
-    );
-
-    const coinType = this.chainGetter.getChain(this.chainId).coinType;
-
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const keplr = (await this.getKeplr())!;
-
-    const signDoc = makeSignDoc(
-      aminoMsgs,
-      fee,
-      this.chainId,
-      memo,
-      account.getAccountNumber().toString(),
-      account.getSequence().toString()
-    );
-
-    const signResponse = await keplr.signAmino(
-      this.chainId,
-      this.bech32Address,
-      signDoc,
-      signOptions
-    );
-
-    const signedTx = this.hasNoLegacyStdFeature()
-      ? cosmos.tx.v1beta1.TxRaw.encode({
-          bodyBytes: cosmos.tx.v1beta1.TxBody.encode({
-            messages: protoMsgs,
-            memo: signResponse.signed.memo,
-          }).finish(),
-          authInfoBytes: cosmos.tx.v1beta1.AuthInfo.encode({
-            signerInfos: [
-              {
-                publicKey: {
-                  type_url:
-                    coinType === 60
-                      ? "/ethermint.crypto.v1.ethsecp256k1.PubKey"
-                      : "/cosmos.crypto.secp256k1.PubKey",
-                  value: cosmos.crypto.secp256k1.PubKey.encode({
-                    key: Buffer.from(
-                      signResponse.signature.pub_key.value,
-                      "base64"
-                    ),
-                  }).finish(),
-                },
-                modeInfo: {
-                  single: {
-                    mode: SignMode.SIGN_MODE_LEGACY_AMINO_JSON,
-                  },
-                },
-                sequence: Long.fromString(signResponse.signed.sequence),
-              },
-            ],
-            fee: {
-              amount: signResponse.signed.fee.amount as ICoin[],
-              gasLimit: Long.fromString(signResponse.signed.fee.gas),
-            },
-          }).finish(),
-          signatures: [Buffer.from(signResponse.signature.signature, "base64")],
-        }).finish()
-      : makeStdTx(signResponse.signed, signResponse.signature);
-
-    return {
-      txHash: await keplr.sendTx(this.chainId, signedTx, mode as BroadcastMode),
-      signDoc: signResponse.signed,
-    };
-  }
-
-  get instance(): AxiosInstance {
-    const chainInfo = this.chainGetter.getChain(this.chainId);
-    return Axios.create({
-      ...{
-        baseURL: chainInfo.rest,
-      },
-      ...chainInfo.restConfig,
-    });
-  }
-
   get walletStatus(): WalletStatus {
     return this._walletStatus;
   }
@@ -537,8 +256,19 @@ export class AccountSetBase<MsgOpts, Queries extends IObject> {
     return this._bech32Address;
   }
 
+  /**
+   * Returns the tx type in progress waiting to be committed.
+   * If there is no tx type in progress, this returns an empty string ("").
+   */
+  get txTypeInProgress(): string {
+    return this._txTypeInProgress;
+  }
+
+  /**
+   * @deprecated Use `txTypeInProgress`
+   */
   get isSendingMsg(): string | boolean {
-    return this._isSendingMsg;
+    return this.txTypeInProgress;
   }
 
   get hasEvmosHexAddress(): boolean {
@@ -548,16 +278,17 @@ export class AccountSetBase<MsgOpts, Queries extends IObject> {
   get evmosHexAddress(): string {
     return evmosToEth(this.bech32Address);
   }
+}
 
-  protected get queries(): DeepReadonly<QueriesSetBase & Queries> {
-    return this.queriesStore.get(this.chainId);
+export class AccountSetBaseSuper extends AccountSetBase {
+  constructor(...params: ConstructorParameters<typeof AccountSetBase>) {
+    super(...params);
+
+    makeObservable(this);
   }
 
-  protected hasNoLegacyStdFeature(): boolean {
-    const chainInfo = this.chainGetter.getChain(this.chainId);
-    return (
-      chainInfo.features != null &&
-      chainInfo.features.includes("no-legacy-stdTx")
-    );
+  @action
+  setTxTypeInProgress(type: string): void {
+    this._txTypeInProgress = type;
   }
 }
